+++ somehal/src/arch/aarch64/entry.rs
@@ -1,20 +1,13 @@
-use crate::{fdt, handle_err, mem::setup_memory_main, println};
+use super::debug;
+use crate::{fdt, handle_err, println};
 
 pub fn mmu_entry() {
-    #[cfg(not(feature = "early-debug"))]
-    arch::Arch::init_debugcon();
+    debug::init();
     println!("MMU ready!");
 
     let cpu_count = handle_err!(fdt::cpu_count(), "could not get cpu count");
 
     println!("{:<12}: {}", "CPU count", cpu_count);
 
-    let memories = handle_err!(fdt::find_memory(), "could not get memories");
-
-    unsafe {
-        setup_memory_main(memories, cpu_count);
-
-
-        
-    }
+    // let memories = handle_err!(fdt::find_memory(), "could not get memories");
 }
